Fix infinite-loop and 32-bit address truncation bugs (#2)
* Fix comma-operator bug in process-switch loop

The comma operator discarded openNextProcess()'s result, so the
condition always evaluated to false and the loop never broke when
no more matching processes were found, risking an infinite spin.

* Fix 32-bit truncation of 64-bit module addresses in offset scanner

NetworkServer::current_offset returned UINT, truncating the QWORD
offset table (e.g. CharInfo at 0x140EB6440) before EQGameScanner
could use the full address as a base for structure-offset scanning.

findEQPointerOffset/findEQStructureOffset took and returned DWORD
despite handling full compile-time module addresses that exceed 32
bits, and callers were explicitly casting QWORD ini values down to
DWORD before passing them in. Widened both functions end-to-end to
QWORD and added proper 8-byte read support for typelen >= 8 (masks
with 8 't' characters previously fell through to a 4-byte DWORD
read, silently discarding the upper 32 bits).

+++ include/EQGameScanner.h
@@ -35,8 +35,8 @@ public:
 public:
 	bool executableExists() const;
 	void setExe(TCHAR* str);
-	DWORD findEQPointerOffset(DWORD startAddress, std::size_t blockSize, const PBYTE byteMask, const PCHAR charMask);
-	DWORD findEQStructureOffset(DWORD startAddress, std::size_t blockSize, const PBYTE byteMask, const PCHAR charMask, const QWORD baseEQPointerAddress);
+	QWORD findEQPointerOffset(QWORD startAddress, std::size_t blockSize, const PBYTE byteMask, const PCHAR charMask);
+	QWORD findEQStructureOffset(QWORD startAddress, std::size_t blockSize, const PBYTE byteMask, const PCHAR charMask, const QWORD baseEQPointerAddress);
 	bool ScanExecutable(HWND hDlg, IniReaderInterface* ir_intf, NetworkServerInterface* net_intf, bool write_out = false);
 	void ScanSecondary(HWND hDlg, IniReaderInterface* ir_intf, NetworkServerInterface* net_intf);
